feat: consolidate audit tests
